refactor: replace movie... components
