+++ src/components/MovieInfomation/MovieParams/MovieParams.module.scss
