+++ src/components/MovieInfomation/MovieParams/MovieParams.tsx
