+++ src/components/MovieInfomation/MovieParams/index.ts
