+++ src/components/MovieInfomation/MovieRating/MovieRating.module.scss
