+++ src/components/MovieInfomation/MovieRating/MovieRating.stories.tsx
