+++ src/components/MovieInfomation/MovieRating/index.ts
